Keeping fixture attribute and test attribute separate.
